requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
adafruit-blinka==6.6.0; python_version >= '3.6'
adafruit-bme280==1.0.1
adafruit-circuitpython-bme280==2.6.1
adafruit-circuitpython-busdevice==5.0.6
adafruit-gpio==1.0.3
adafruit-platformdetect==3.8.0; python_version >= '3.6'
adafruit-pureio==1.1.8; python_version >= '3.5'
board==1.0
colorlog==5.0.1
greenlet==1.0.0; python_version >= '3'
marshmallow-sqlalchemy==0.24.3
marshmallow==3.11.1; python_version >= '3.5'
paho-mqtt==1.5.1
psycopg2-binary==2.8.6
pyftdi==0.52.9; python_version >= '3.5'
pyserial==3.5
python-dotenv==0.17.1
pyusb==1.1.1
spidev==3.5
sqlalchemy==1.4.12
